requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-doc
annotated-types
anyio
attrs
backoff
bcrypt
beautifulsoup4
blinker
blis
build
cachetools
catalogue
certifi
cffi
charset-normalizer
chromadb
click
cloudpathlib
confection
cryptography
cymem
dataclasses-json
distro
dnspython
durationpy
ecdsa
email-validator
emoji
faiss-cpu
faker
fastapi
ffmpeg-python
filelock
filetype
flatbuffers
fpdf
frozenlist
fsspec
future
gitdb
gitpython
googleapis-common-protos
grpcio
h11
hf-xet
html5lib
httpcore
httptools
httpx
httpx-sse
huggingface-hub
idna
importlib-metadata
importlib-resources
installer
jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-classic
langchain-community
langchain-core
langchain-text-splitters
langdetect
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
llvmlite
lxml
markdown-it-py
markupsafe
marshmallow
mdurl
mmh3
more-itertools
mpmath
multidict
murmurhash
mypy-extensions
narwhals
networkx
numba
numpy
oauthlib
olefile
onnxruntime
openai
openai-whisper
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
passlib
pdfminer-six
pdfplumber
pillow
pinecone-client
pinecone-plugin-interface
posthog
preshed
propcache
protobuf
psutil
psycopg2-binary
pyarrow
pyasn1
pybase64
pycparser
pydantic
pydantic-core
pydantic-settings
pydeck
pygments
pypdf
pypdfium2
pypika
pyproject-hooks
python-dateutil
python-docx
python-dotenv
python-iso639
python-jose
python-magic
python-oxmsg
pytz
pyyaml
rank-bm25
rapidfuzz
redis
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smart-open
smmap
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
sqlalchemy
srsly
starlette
streamlit
sympy
tenacity
thinc
threadpoolctl
tiktoken
tokenizers
toml
torch
torchaudio
torchvision
tornado
tqdm
transformers
typer
typer-slim
typing-extensions
typing-inspect
typing-inspection
tzdata
unstructured
unstructured-client
urllib3
uuid-utils
uvicorn
uvloop
wasabi
watchfiles
weasel
webencodings
websocket-client
websockets
wrapt
xxhash
yarl
zipp
zstandard
